combined_rag_api/requirements.txt
fastapi
uvicorn
asyncpg
uuid
pydantic
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=1.3.0
openpyxl>=3.0.9
beautifulsoup4>=4.10.0
aiohttp>=3.8.1
trafilatura>=1.4.0
markdown>=3.3.6
sentence-transformers>=2.2.2
numpy>=1.21.0
